**Summary:** Similar to https://github.com/pytorch/ao/pull/2986,
this commit improves the prepare vs convert SQNR of NVFP4 QAT
from 12 to 36 with `use_per_tensor_scale`, and 12 to inf without.
This is achieved by mimicking the PTQ flow more closely,
in particular, in descending order of significance:

1. Simulate `f4_unpacked_to_f32` and `f32_to_f4_unpacked`,
   but in `torch.int32` instead of `torch.uint8`
2. Do not cast intermediate fake quantized values to original
   dtype, e.g. bf16 which loses some fidelity from fp32
3. Fake round blockwise scales to float8

**Test Plan:**
```
python test/quantization/test_qat.py -k test_qat_nvfp4
python test/quantization/test_qat.py -k test_quantize_api_nvfp4
```

End-to-end tests TBD.

[ghstack-poisoned]

+++ torchao/prototype/qat/nvfp4.py
@@ -27,9 +27,7 @@ class _FP4Round(torch.autograd.Function):
 
     @staticmethod
     def forward(ctx, x: torch.Tensor) -> torch.Tensor:
-        q = _f32_to_floatx_unpacked(
-            x, EBITS_F4_E2M1, MBITS_F4_E2M1, compute_dtype=torch.int32
-        )
+        q = _f32_to_floatx_unpacked(x, EBITS_F4_E2M1, MBITS_F4_E2M1)
         dq = _floatx_unpacked_to_f32(q, EBITS_F4_E2M1, MBITS_F4_E2M1)
         return dq
